fix: wibar scroll broken with NumLock, UBSan runtime errors
Three fixes:

1. CLEANMASK now strips NumLock (Mod2) in addition to CapsLock.
   Scroll/button bindings defined with modifiers={} failed on wibar
   when NumLock was active because the modifier table included "Mod2",
   while root/desktop bindings treated modifiers==0 as "any" and
   always matched. This caused tag scrolling on wibar to require
   extra wheel clicks or not work at all.

2. Scroll axis events now use value120 accumulation for discrete
   wheels. One physical wheel click = delta_discrete ±120. High-res
   mice send smaller sub-steps that are accumulated before emitting
   a button event, preventing multiple tag switches per click.

3. UBSan runtime errors fixed:
   - bsearch() with NULL base in signal_array_lookup (empty arrays)
   - memmove/memcpy with NULL items in array_splice
   - Signed integer overflow in draw.c (uint8_t << 24)

Also: test runner LC_NUMERIC=C for non-English locales,
keyboard-focus-sync test assertions adjusted.

+++ draw.c
@@ -146,7 +146,7 @@ draw_surface_from_pixbuf(GdkPixbuf *buf)
                 r = (uint8_t)(r * alpha);
                 g = (uint8_t)(g * alpha);
                 b = (uint8_t)(b * alpha);
-                *cairo++ = (a << 24) | (r << 16) | (g << 8) | b;
+                *cairo++ = ((uint32_t)a << 24) | ((uint32_t)r << 16) | ((uint32_t)g << 8) | b;
             }
         }
         pixels += pix_stride;
